requirements-app.txt
# Our in-house packagesbp-accessible-output2==0.5bp-db-magic==0.2bp-lib-loader==0.21bp-platform-utils==0.40bp-winpaths==0.2;sys_platform=="win32"# Our fork of the configobj (py3k compatible version).https://github.com/blindpandas/configobj/tarball/v5.1.0.dev0attrs==21.2.0babel==2.9.1banal==1.0.4beautifulsoup4==4.9.3blinker==1.4chemical==1.0.1demandimport==0.3.4diskcache==5.2.1docrpy==0.1.1;sys_platform=="win32"ebooklib==0.17.1ftfy==6.0.3inscriptis==2.0.0languagecodes==1.0.7lazy_import==0.2.2lru-dict==1.1.7lxml==4.6.3mammoth==1.4.17mistune==0.8.4mobi==0.3.1more-itertools==8.7.0msgpack==1.0.2numpy==1.20.0odfpy==1.4.1opencv-python-headless==4.5.1.48peewee==3.14.8Pillow==8.1.0pycld2==0.41pydantic==1.8.1PyMuPDF==1.19.2pytesseract==0.3.8python-docx==0.8.10pythonnet==2.5.2;sys_platform=="win32"python-pptx==0.6.18pywin32==228;sys_platform=="win32"pyxpdf==0.2.3pyxpdf-data==1.1.0trafilatura==0.9.2regex==2021.9.30requests==2.25.1selectolax==0.3.6SQLAlchemy==1.3.24ujson==4.0.2unicode-slugify==0.1.3url-normalize==1.4.3wikipedia==1.4.0wxPython==4.1.1yarl==1.7.2